refactor: move Go backend code from tun2socks to app
