+++ Android/app/src/main/go/intra/android/init.go
