+++ Android/app/src/main/go/intra/android/tun2socks.go
@@ -22,9 +22,9 @@ import (
 	"os"
 	"strings"
 
-	"github.com/Jigsaw-Code/Intra/Android/tun2socks/intra"
-	"github.com/Jigsaw-Code/Intra/Android/tun2socks/intra/doh"
-	"github.com/Jigsaw-Code/Intra/Android/tun2socks/intra/protect"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra/doh"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra/protect"
 	"github.com/Jigsaw-Code/outline-sdk/network"
 )
