+++ Android/app/src/main/go/intra/doh/atomic.go
