+++ Android/app/src/main/go/intra/split/direct_split.go
