+++ Android/app/src/main/go/intra/split/retrier.go
