+++ Android/app/src/main/go/intra/tcp.go
@@ -23,7 +23,7 @@ import (
 	"sync/atomic"
 	"time"
 
-	"github.com/Jigsaw-Code/Intra/Android/tun2socks/intra/split"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra/split"
 	"github.com/Jigsaw-Code/outline-sdk/transport"
 )
