+++ Android/app/src/main/go/intra/tunnel.go
@@ -22,8 +22,8 @@ import (
 	"os"
 	"strings"
 
-	"github.com/Jigsaw-Code/Intra/Android/tun2socks/intra/doh"
-	"github.com/Jigsaw-Code/Intra/Android/tun2socks/intra/protect"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra/doh"
+	"github.com/Jigsaw-Code/Intra/Android/app/src/main/go/intra/protect"
 	"github.com/Jigsaw-Code/outline-sdk/network"
 	"github.com/Jigsaw-Code/outline-sdk/network/lwip2transport"
 )
