+++ Android/gradle.properties
@@ -24,3 +24,5 @@ ANDROID_BUILD_TOOLS_VERSION=33.0.0
 
 ANDROID_MIN_SDK_VERSION=16
 ANDROID_TARGET_SDK_VERSION=33
+
+org.gradle.jvmargs=-Xmx2048M
