--- Android/tun2socks/src/main/AndroidManifest.xml
@@ -1,3 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest package="app.intra.tun2socks">
-</manifest>
